+++ MouseDodge/MouseDodge/DynamicMonster.h
@@ -6,18 +6,18 @@ class BaseMonster;
 class DynamicMonster : public BaseMonster
 {
 public:
-	DynamicMonster(double sizeSpeed = 1, double speedX = 1, double speedY = 1, float radius = 10, float posX = 0, float posY = 0) :
+	DynamicMonster(float sizeSpeed = 1, float speedX = 1, float speedY = 1, float radius = 10, float posX = 0, float posY = 0) :
 		BaseMonster(speedX, speedY, radius, posX, posY)
 	{
 		Init(sizeSpeed);
 	}
 	~DynamicMonster();
 
-	void Init(double sizeSpeed);
+	void Init(float sizeSpeed);
 	void Update(sf::RenderWindow& window);
 
 private:
-	double sizeSpeed;
+	float sizeSpeed;
 	int count;
 	bool increasing;
 };
